requirements.txt
Pillow
opencv-python
cymysql
